+++ client/app/components/RatingsAndReviews/RatingsAndReviews.css
@@ -58,8 +58,10 @@
   border: 3px solid #333;
   background-color: #dfdfdf;
   min-height: 100px;
+  max-height: 100%;
   margin: 10px;
 }
+
 
 .reviewButtonsContainer {
   display: flex;
