+++ client/app/components/RatingsAndReviews/ReviewList/ReviewButtonsContainer/AddReviewButton/ReviewCharacteristics/ReviewCharacteristic.jsx
@@ -1,5 +1,6 @@
 /* eslint-disable react/prop-types */
 import React, { useState } from 'react';
+import { getCharacteristicDescriptions } from '../../../../../helperFunctions/helperFunctions.js';
 
 const ReviewCharacteristic = ({ name, id, reviewCharsObj, setReviewCharsObj }) => {
   const [rowCharSelection, setRowCharSelection] = useState(0);
@@ -11,50 +12,50 @@ const ReviewCharacteristic = ({ name, id, reviewCharsObj, setReviewCharsObj }) =
     setReviewCharsObj({ ...reviewCharsObj, [id]: Number(e.target.value) });
   };
 
-  const reviewCharacteristicsObject = {};
+  const reviewCharacteristicsObject = getCharacteristicDescriptions(name);
 
-  if (name === 'Size') {
-    reviewCharacteristicsObject.one = 'A size too small';
-    reviewCharacteristicsObject.two = '½ a size too small';
-    reviewCharacteristicsObject.three = 'Perfect';
-    reviewCharacteristicsObject.four = '½ a size too big';
-    reviewCharacteristicsObject.five = 'A size too wide';
-  }
-  if (name === 'Width') {
-    reviewCharacteristicsObject.one = 'Too narrow';
-    reviewCharacteristicsObject.two = 'Slightly narrow';
-    reviewCharacteristicsObject.three = 'Perfect';
-    reviewCharacteristicsObject.four = 'Slightly wide';
-    reviewCharacteristicsObject.five = 'Too wide';
-  }
-  if (name === 'Comfort') {
-    reviewCharacteristicsObject.one = 'Uncomfortable';
-    reviewCharacteristicsObject.two = 'Slightly uncomfortable';
-    reviewCharacteristicsObject.three = 'Ok';
-    reviewCharacteristicsObject.four = 'Comfortable';
-    reviewCharacteristicsObject.five = 'Perfect';
-  }
-  if (name === 'Quality') {
-    reviewCharacteristicsObject.one = 'Poor';
-    reviewCharacteristicsObject.two = 'Below average';
-    reviewCharacteristicsObject.three = 'What I expected';
-    reviewCharacteristicsObject.four = 'Pretty great';
-    reviewCharacteristicsObject.five = 'Perfect';
-  }
-  if (name === 'Length') {
-    reviewCharacteristicsObject.one = 'Runs short';
-    reviewCharacteristicsObject.two = 'Runs slightly short';
-    reviewCharacteristicsObject.three = 'Perfect';
-    reviewCharacteristicsObject.four = 'Runs slightly long';
-    reviewCharacteristicsObject.five = 'Runs long';
-  }
-  if (name === 'Fit') {
-    reviewCharacteristicsObject.one = 'Runs tight';
-    reviewCharacteristicsObject.two = 'Runs slightly tight';
-    reviewCharacteristicsObject.three = 'Perfect';
-    reviewCharacteristicsObject.four = 'Runs slightly long';
-    reviewCharacteristicsObject.five = 'Runs long';
-  }
+  // if (name === 'Size') {
+  //   reviewCharacteristicsObject.one = 'A size too small';
+  //   reviewCharacteristicsObject.two = '½ a size too small';
+  //   reviewCharacteristicsObject.three = 'Perfect';
+  //   reviewCharacteristicsObject.four = '½ a size too big';
+  //   reviewCharacteristicsObject.five = 'A size too wide';
+  // }
+  // if (name === 'Width') {
+  //   reviewCharacteristicsObject.one = 'Too narrow';
+  //   reviewCharacteristicsObject.two = 'Slightly narrow';
+  //   reviewCharacteristicsObject.three = 'Perfect';
+  //   reviewCharacteristicsObject.four = 'Slightly wide';
+  //   reviewCharacteristicsObject.five = 'Too wide';
+  // }
+  // if (name === 'Comfort') {
+  //   reviewCharacteristicsObject.one = 'Uncomfortable';
+  //   reviewCharacteristicsObject.two = 'Slightly uncomfortable';
+  //   reviewCharacteristicsObject.three = 'Ok';
+  //   reviewCharacteristicsObject.four = 'Comfortable';
+  //   reviewCharacteristicsObject.five = 'Perfect';
+  // }
+  // if (name === 'Quality') {
+  //   reviewCharacteristicsObject.one = 'Poor';
+  //   reviewCharacteristicsObject.two = 'Below average';
+  //   reviewCharacteristicsObject.three = 'What I expected';
+  //   reviewCharacteristicsObject.four = 'Pretty great';
+  //   reviewCharacteristicsObject.five = 'Perfect';
+  // }
+  // if (name === 'Length') {
+  //   reviewCharacteristicsObject.one = 'Runs short';
+  //   reviewCharacteristicsObject.two = 'Runs slightly short';
+  //   reviewCharacteristicsObject.three = 'Perfect';
+  //   reviewCharacteristicsObject.four = 'Runs slightly long';
+  //   reviewCharacteristicsObject.five = 'Runs long';
+  // }
+  // if (name === 'Fit') {
+  //   reviewCharacteristicsObject.one = 'Runs tight';
+  //   reviewCharacteristicsObject.two = 'Runs slightly tight';
+  //   reviewCharacteristicsObject.three = 'Perfect';
+  //   reviewCharacteristicsObject.four = 'Runs slightly long';
+  //   reviewCharacteristicsObject.five = 'Runs long';
+  // }
 
   const CharExplanationRender = () => {
     if (rowCharSelection !== 0) {
